增加内容，修复 1.增加annotation包，用于标记mapper方法，便于输出日志 2.增加aspect包，用于编写切面，完成简单日志 3.为DTO类和model类重写toString() 4.为RailwayService修改deleteLine()和deleteStation()逻辑，避免空值，并优化了数据库查询次数

+++ src/main/java/com/RailManager/demo/service/RailwayService.java
@@ -3,6 +3,7 @@ package com.RailManager.demo.service;
 import com.RailManager.demo.DTO.LineDTO;
 import com.RailManager.demo.DTO.LineInfoDTO;
 import com.RailManager.demo.DTO.StationDTO;
+import com.RailManager.demo.annotation.MyService;
 import com.RailManager.demo.mapper.LineMapper;
 import com.RailManager.demo.mapper.StationMapper;
 import com.RailManager.demo.model.Line;
@@ -19,42 +20,25 @@ public class RailwayService {
     LineMapper lineMapper;
     @Autowired
     StationMapper stationMapper;
-    //按名字获取Line
-    private Line getLineByName(String lineName){
-        return lineMapper.getLineByName(lineName);
-    }
-    //获取所有Line
-    private List<Line> getAllLines(){
-        return lineMapper.getAllLines();
-    }
-    //按ID获取站点
-    private Station getStationById(Integer stationId){
-        return stationMapper.getStationById(stationId);
-    }
-    //获取全部站点
-    private List<Station> getAllStations(){
-        return stationMapper.getAllStations();
-    }
-    //按线名获取Station
-    private List<Station> getStationsByLine(String lineName){
-        return stationMapper.getStationsByLine(lineName);
-    }
+    @MyService
     //获取某线DTO
     public LineInfoDTO getLineInfoDTOByName(String lineName){
         LineInfoDTO dto = new LineInfoDTO();
-        dto.setLine(getLineByName(lineName));
-        dto.setStations(getStationsByLine(lineName));
+        dto.setLine(lineMapper.getLineByName(lineName));
+        dto.setStations(stationMapper.getStationsByLine(lineName));
         return dto;
     }
+    @MyService
     //获取全部线DTO
     public List<LineInfoDTO> getAllLineInfoDTO(){
         List<LineInfoDTO> dtos = new ArrayList<>();
-        List<Line> lines = getAllLines();
+        List<Line> lines = lineMapper.getAllLines();
         for(Line line : lines) {
             dtos.add(getLineInfoDTOByName(line.getLineName()));
         }
         return dtos;
     }
+    @MyService
     //插入新线信息:线名相同为更新
     public ResponseEntity<String> addLine(LineDTO dto){
         Line line = new Line(dto);
@@ -72,6 +56,7 @@ public class RailwayService {
             }
         }
     }
+    @MyService
     //插入新站信息：需要获取stationId 获取和后移innerId 修改对应线的stationNum
     public ResponseEntity<String> addStation(StationDTO dto){
         Station station = new Station();
@@ -86,11 +71,11 @@ public class RailwayService {
         //修改innerId
         Integer preId = dto.getPreStationId();
         Integer nextId = dto.getNextStationId();
+        List<Station> stationsByLine = stationMapper.getStationsByLine(dto.getLineName());
         if(nextId == 0){
             station.setInnerId(preId + 1);
         }
         else {
-            List<Station> stationsByLine = getStationsByLine(dto.getLineName());
             for(Station st : stationsByLine){
                 if(preId < st.getInnerId()){
                     stationMapper.updateInnerId(st.getStationId(), st.getInnerId() + 1);
@@ -100,46 +85,54 @@ public class RailwayService {
         }
         stationMapper.insertStation(station);
         //修改Line.stationNum
-        lineMapper.updateStationNum(dto.getLineName(), getStationsByLine(dto.getLineName()).size());
+        lineMapper.updateStationNum(dto.getLineName(), stationsByLine.size() + 1);
 
         return ResponseEntity.ok("Station inserted.");
     }
+    @MyService
     //删除线信息 级联删除站信息
     public ResponseEntity<String> deleteLine(String lineName){
+        //判断被删除线路是否存在
+        Line delLine = lineMapper.getLineByName(lineName);
+        if(delLine == null) return ResponseEntity.ok("No need to delete.");
+        //不存在则删除
+        lineMapper.deleteLine(lineName);
+        //级联删除站点
         List<Station> stations = stationMapper.getStationsByLine(lineName);
         if(!stations.isEmpty()){
             for(Station st : stations) {
                 deleteStation(st.getStationId());
             }
         }
-        if(lineMapper.deleteLine(lineName) != 0) {
-            return ResponseEntity.ok("Line deleted.");
-        }
-        else return ResponseEntity.ok("No need to delete.");
+        return ResponseEntity.ok("Line deleted.");
     }
+    @MyService
     //删除站信息 需要前移stationId 前移innerId 修改对应线的stationNum
     public ResponseEntity<String> deleteStation(Integer stationId) {
-        String lineName = getStationById(stationId).getLineName();
-        if (stationMapper.deleteStation(stationId) != 0) {
-            List<Station> stations = stationMapper.getAllStations();
-            //处理stationId
-            for (Station station : stations) {
-                if (stationId < station.getStationId()) {
-                    stationMapper.updateStationId(station.getStationId(), station.getStationId() - 1);
-                }
+        //判断被删除站点是否存在
+        Station delStation = stationMapper.getStationById(stationId);
+        if(delStation == null) return ResponseEntity.ok("No need to delete.");
+        //存在则作后续处理并删除
+        stationMapper.deleteStation(stationId);
+        String lineName = delStation.getLineName();
+        //处理innerId
+        List<Station> stationsByLine = stationMapper.getStationsByLine(lineName);
+        Integer innerId = delStation.getInnerId();
+        for (Station st : stationsByLine) {
+            if (innerId < st.getInnerId()) {
+                stationMapper.updateInnerId(st.getStationId(), st.getInnerId() - 1);
             }
-            //处理innerId
-            List<Station> stationsByLine = stationMapper.getStationsByLine(lineName);
-            for (Station st : stationsByLine) {
-                if (getStationById(stationId).getInnerId() < st.getInnerId()) {
-                    stationMapper.updateInnerId(st.getStationId(), st.getInnerId() - 1);
-                }
-            }
-            //修改Line.stationNum
-            lineMapper.updateStationNum(lineName, getStationsByLine(lineName).size());
-
-            return ResponseEntity.ok("Station deleted.");
         }
-        else return ResponseEntity.ok("No need to delete.");
+        //处理stationId
+        List<Station> stations = stationMapper.getAllStations();
+        for (Station station : stations) {
+            if (stationId < station.getStationId()) {
+                stationMapper.updateStationId(station.getStationId(), station.getStationId() - 1);
+            }
+        }
+        //修改Line.stationNum
+        lineMapper.updateStationNum(lineName,  stationsByLine.size() - 1);
+
+        return ResponseEntity.ok("Station deleted.");
     }
 }
